refactor : '숙련도'를 나타내는 class명 변경 및 적용(Skill -> Proficiency)
